% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_fusion_calls.R
\name{aggregate_fusion_calls}
\alias{aggregate_fusion_calls}
\title{Function to aggregate Caller and read support per fusions calls}
\usage{
aggregate_fusion_calls(standardFusioncalls = standardFusioncalls)
}
\arguments{
\item{standardFusioncalls}{A dataframe from star fusion or arriba standardized to run through the filtering steps}
}
\value{
Standardized fusion calls with aggregated Caller and read support
}
\description{
Function to aggregate Caller and read support per fusions calls
}
