% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annoFuseSingleSample.R
\name{annoFuseSingleSample}
\alias{annoFuseSingleSample}
\title{Single Sample use for annoFuse
Performs artifact filter to remove readthroughs,red flags and performs expression filtering with user provided expression Matrix and expression threshold}
\usage{
annoFuseSingleSample(fusionfileArriba = fusionfileArriba,
  fusionfileStarFusion = fusionfileStarFusion,
  expressionFile = expressionFile, expressionFilter = 1,
  tumorID = "tumorID",
  artifactFilter = "GTEx_Recurrent|DGD_PARALOGS|Normal|BodyMap|ConjoinG",
  readingFrameFilter = "in-frame|frameshift|other",
  junctionReadCountFilter = 1, spanningFragCountFilter = 10,
  readthroughFilter = FALSE)
}
\arguments{
\item{fusionfileArriba}{A dataframe from arriba fusion caller}

\item{fusionfileStarFusion}{A dataframe from starfusion caller}

\item{expressionFile}{Expression matrix for samples used in cohort for fusion calls}

\item{expressionFilter}{FPKM/TPM threshold for not expressed}

\item{tumorID}{Sample name to be used}

\item{artifactFilter}{A red flag filter from Annotation ; in OpenPBTA annotation is from FusionAnnotator column "annots"}

\item{readingFrameFilter}{A regex to capture readingframe (eg. in-frame|frameshift|other)}

\item{junctionReadCountFilter}{An integer threshold for JunctionReadCount}

\item{spanningFragCountFilter}{An integer threshold for (SpanningFragCount - JunctionReadCount)}

\item{readthroughFilter}{Boolean for filtering readthroughs}
}
\value{
Standardized fusion calls annotated with gene list and fusion list provided in reference folder
}
\description{
Single Sample use for annoFuse
Performs artifact filter to remove readthroughs,red flags and performs expression filtering with user provided expression Matrix and expression threshold
}
