% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_fusion_calls.R
\name{annotate_fusion_calls}
\alias{annotate_fusion_calls}
\title{Function to annotate fusion calls}
\usage{
annotate_fusion_calls(
  standardFusioncalls = standardFusioncalls,
  geneListReferenceDataTab = geneListReferenceDataTab,
  fusionReferenceDataTab = fusionReferenceDataTab
)
}
\arguments{
\item{standardFusioncalls}{A dataframe from star fusion or arriba standardized to run through the filtering steps}

\item{geneListReferenceDataTab}{A dataframe with column 1 as GeneName 2 source file 3 type; collapse to summarize type}

\item{fusionReferenceDataTab}{A dataframe with column 1 as FusionName 2 source file 3 type; collapse to summarize type}
}
\value{
Standardized fusion calls annotated with gene list and fusion list provided in reference folder
}
\description{
Function to annotate fusion calls
}
\examples{
# TODOTODO

}
