% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/called_by_n_callers.R
\name{called_by_n_callers}
\alias{called_by_n_callers}
\title{Function to identify fusions called by at least n callers}
\usage{
called_by_n_callers(
  standardFusioncalls = standardFusioncalls,
  numCaller = numCaller
)
}
\arguments{
\item{standardFusioncalls}{A dataframe from star fusion or arriba standardized to run through the filtering steps}

\item{numCaller}{Least number of callers that have the fusion}
}
\value{
Fusions called by n callers
}
\description{
Function to identify fusions called by at least n callers
}
\examples{
# TODOTODO

}
