% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression_filter_fusion.R
\name{expression_filter_fusion}
\alias{expression_filter_fusion}
\title{Expression filtering with user provided expression Matrix and standard fusion calls}
\usage{
expression_filter_fusion(
  standardFusioncalls,
  expressionMatrix,
  expressionFilter
)
}
\arguments{
\item{standardFusioncalls}{A dataframe from star fusion or arriba standardized to run through the filtering steps}

\item{expressionMatrix}{Expression matrix for samples used in cohort for fusion calls}

\item{expressionFilter}{FPKM/TPM threshold for not expressed}
}
\value{
Standardized fusion calls annotated with gene list and fusion list provided in reference folder
}
\description{
Expression filtering with user provided expression Matrix and standard fusion calls
}
\examples{
out_annofuse <- system.file("extdata", "PutativeDriverAnnoFuse_test_v14.tsv", package = "annoFuse")
sfc <- read.delim(out_annofuse)
# TODOTODO: what are some good values for expressionMatrix and expressionFilter?

}
