% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression_filtering.R
\name{expressionFilterFusion}
\alias{expressionFilterFusion}
\title{Expression filtering with user provided expression Matrix and standard fusion calls}
\usage{
expressionFilterFusion(
  standardFusioncalls = standardFusioncalls,
  expressionMatrix = expressionMatrix,
  expressionFilter = expressionFilter
)
}
\arguments{
\item{standardFusioncalls}{A dataframe from star fusion or arriba standardized to run through the filtering steps}

\item{expressionMatrix}{Expression matrix for samples used in cohort for fusion calls}

\item{expressionFilter}{FPKM/TPM threshold for not expressed}
}
\value{
Standardized fusion calls annotated with gene list and fusion list provided in reference folder
}
\description{
Expression filtering with user provided expression Matrix and standard fusion calls
}
