% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filt_exp.R
\name{filt_exp}
\alias{filt_exp}
\title{Filter fusion calls with 5' and 3' genes expressed
Filtering for expresion of 3' and 5' allows us to look for relevant fusions}
\usage{
filt_exp(fusion_calls = fusion_calls, expdata = expdata, clin = clin)
}
\arguments{
\item{fusion_calls}{data frame of fusion calls with histologies}

\item{expdata}{tsv file of rsem data}

\item{clin}{clinical file to annotate}
}
\value{
filtered fusion calls
}
\description{
Filter fusion calls with 5' and 3' genes expressed
Filtering for expresion of 3' and 5' allows us to look for relevant fusions
}
\examples{
fusion_calls<-system.file("extdata","filt_fusion_calls.RData",package="annofusion")
rsem_file<-system.file("extdata","test_rsem.RData",package="annofusion")
clin<-system.file("extdata","2019-06-07-pbta-histologies.tsv",package="annofusion")
load(fusion_calls)
load(rsem_file)
clin<-read.delim(clin,sep="\\t",stringsAsFactors=FALSE)
fusion_calls<-filt_fusion_calls$driver
filt_exp(fusion_calls=fusion_calls,expdata=expdata,clin=clin)
}
