% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filt_prioritization.R
\name{filt_prioritization}
\alias{filt_prioritization}
\title{Filter for putative driver fusions}
\usage{
filt_prioritization(fusion_calls = fusion_calls,
  lit_fusions = lit_fusions, tsgs = tsgs, onco = onco, tcga = tcga,
  remove_circularRNA = TRUE)
}
\arguments{
\item{fusion_calls}{A dataframe with fusion calls and broad histologies}

\item{lit_fusions}{table of driver fusiions per histologies}

\item{tsgs}{dataframe for tumor suppressor gene list}

\item{onco}{dataframe for oncogenic gene list}

\item{tcga}{dataframe for tcga fusion}

\item{remove_circularRNA}{boolean}
}
\value{
filtered putative fusion calls and annotation
}
\description{
Users can use a driver gene and fusion list or biologically relevant gene
sets like tumor suppersor genes, oncogenic genes, tcga fusions
}
\examples{
fusion_calls<-system.file("extdata","filt_read.RData",package="annofusion")
load(fusion_calls)
###optional##
lit_file<-system.file("extdata", "driver-fusions-v4.txt", package = "annofusion")
lit_fusions<-read.delim(lit_file,stringsAsFactors=FALSE)
tsgs_file<-system.file("extdata", "tsgs.txt", package = "annofusion")
tsgs<-read.delim(tsgs_file,stringsAsFactors=FALSE)
onco_file<-system.file("extdata", "oncogenes.txt", package = "annofusion")
onco<-read.delim(onco_file,stringsAsFactors=FALSE)
tcga_file<-system.file("extdata", "tcga-fusions.txt", package = "annofusion")
tcga<-read.delim(tcga_file,stringsAsFactors=FALSE)
filt_prioritization(fusion_calls=fusion_calls, tsgs=tsgs,onco=onco,tcga=tcga,remove_circularRNA=TRUE)



}
