% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filt_read.R
\name{filt_read}
\alias{filt_read}
\title{Filter fusion calls for read depth, frame information and artifacts
Inframe and frameshift fusions with high supporting reads are considered true calls for further annotation}
\usage{
filt_read(fusion_calls = fusion_calls, read_support = read_support,
  caller = caller, remove_artifacts = TRUE)
}
\arguments{
\item{fusion_calls}{A dataframe from star fusion or arriba (more callers to be added )}

\item{read_support}{An integer for spanning read filter}

\item{caller}{string options STARfusion/arriba}

\item{remove_artifacts}{boolean}
}
\value{
Inframe and frameshift filtered fusion calls for user provided read depth
}
\description{
Filter fusion calls for read depth, frame information and artifacts
Inframe and frameshift fusions with high supporting reads are considered true calls for further annotation
}
\examples{
fusion_file<-system.file("extdata", "test_star.tsv", package = "annofusion")
fusion_calls<-read.delim(fusion_file,stringsAsFactors=FALSE)
read_support<-10
caller<-"STARfusion"
filt_read(fusion_calls=fusion_calls,read_support=read_support, caller=caller)


}
