% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filt_anno.R
\name{fusion_anno}
\alias{fusion_anno}
\title{Fusion annotation with tsgs oncogenes kinase tcga fusion}
\usage{
fusion_anno(fusion_calls = filt_fusion_exp_genes, clin = clin,
  tsgs = tsgs, onco = onco, tcga = tcga, kinase = kinase,
  chimerDB = chimerDB, tf = tf)
}
\arguments{
\item{fusion_calls}{list of putative drivers and total filtered fusion calls}

\item{clin}{clinical info}

\item{tsgs}{tsgs gene list}

\item{onco}{oncogenic gene list}

\item{tcga}{tcga fusion list}

\item{kinase}{kinase gene list}

\item{chimerDB}{chimerDB}

\item{tf}{transcription factor gene list}
}
\value{
filtered annotated fusion calls
}
\description{
Fusion annotation with tsgs oncogenes kinase tcga fusion
}
\examples{
driver_fusions<-system.file("extdata","fusion_exp_filt.Rdata",package="annofusion")
prioritized_fusions<-system.file("extdata","fusion_exp_filt_prioritized.Rdata",package="annofusion")
load(driver_fusions)
load(prioritized_fusions)
filt_fusion_exp_genes<-list("driver"=driver,"prioritized"=prioritized)
tsgs_file<-system.file("extdata", "tsgs.txt", package = "annofusion")
tsgs<-read.delim(tsgs_file,stringsAsFactors=FALSE)
onco_file<-system.file("extdata", "oncogenes.txt", package = "annofusion")
onco<-read.delim(onco_file,stringsAsFactors=FALSE)
tcga_file<-system.file("extdata", "tcga-fusions.txt", package = "annofusion")
tcga<-read.delim(tcga_file,stringsAsFactors=FALSE)
kinase_file<-system.file("extdata", "Kincat_Hsap.08.02.txt", package = "annofusion")
kinase<-read.delim(kinase_file,stringsAsFactors=FALSE)
chimerDB_file<-system.file("extdata","ChimerDB3.0_ChimerKB.txt", package = "annofusion")
chimerDB<-read.delim(chimerDB_file)
tf_file<-system.file("extdata","TRANSFAC_TF.txt", package = "annofusion")
tf<-read.delim(tf_file,header=FALSE,stringsAsFactor=FALSE)
clin_file<-system.file("extdata","2019-06-07-pbta-histologies.tsv",package="annofusion")
clin<-read.delim(clin_file,sep="\\t",stringsAsFactors=FALSE)
clin$Sample<-paste(clin$sample_id,clin$aliquot_id,sep="_")
fusion_anno(fusion_calls=filt_fusion_exp_genes, tsgs=tsgs,onco=onco,tcga=tcga,kinase=kinase,chimerDB=chimerDB,tf=tf,clin=clin)

}
