% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusion_filtering_QC.R
\name{fusion_filtering_QC}
\alias{fusion_filtering_QC}
\title{Filters standardized fusion calls to remove artifacts and false positives.}
\usage{
fusion_filtering_QC(
  standardFusioncalls,
  readingFrameFilter = "in-frame|frameshift|other",
  artifactFilter,
  junctionReadCountFilter,
  spanningFragCountFilter,
  readthroughFilter = TRUE
)
}
\arguments{
\item{standardFusioncalls}{A dataframe from star fusion or arriba standardized to run through the filtering steps}

\item{readingFrameFilter}{A regex to capture readingframe (eg. in-frame|frameshift|other)}

\item{artifactFilter}{A red flag filter from Annotation ; in OpenPBTA annotation is from FusionAnnotator column "annots"}

\item{junctionReadCountFilter}{An integer threshold for JunctionReadCount}

\item{spanningFragCountFilter}{An integer threshold for (SpanningFragCount - JunctionReadCount)}

\item{readthroughFilter}{Boolean for filtering readthroughs}
}
\value{
Standardized fusion calls filtered to pass QC and remove calls with insufficient read-support and annotation red-flags
}
\description{
Events such as polymerase read-throughs, mis-mapping due to gene homology, and fusions occurring in healthy normal
tissue require stringent filtering, making it difficult for researchers and clinicians to discern true underlying
oncogenic drivers of a tumor and in some cases, appropriate therapy
}
\examples{
# standardize
fusionfileArriba <- read.delim(
  system.file("extdata", "arriba_example.tsv", package = "annoFuse"), stringsAsFactors = FALSE)
fusionfileStarFusion <- read.delim(
  system.file("extdata", "starfusion_example.tsv", package = "annoFuse"), stringsAsFactors = FALSE)
formattedArriba <- fusion_standardization(fusionfileArriba,
                                          caller = "ARRIBA",
                                          tumorID = "tumorID")
formattedStarFusion <- fusion_standardization(fusionfileStarFusion,
                                              caller="STARFUSION",
                                              tumorID = "tumorID")
# merge standardized fusion calls
standardFusioncalls <- as.data.frame(rbind(formattedStarFusion, formattedArriba))
fusionQCFiltered <- fusion_filtering_QC(
  standardFusioncalls = standardFusioncalls, 
  readingFrameFilter = "in-frame|frameshift|other",
  artifactFilter = "GTEx_Recurrent|DGD_PARALOGS|Normal|BodyMap|ConjoinG",
  junctionReadCountFilter = 1,
  spanningFragCountFilter = 10,
  readthroughFilter = TRUE)
}
