% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusion_multifused.R
\name{fusion_multifused}
\alias{fusion_multifused}
\title{Function to identify fusions called by at least n callers}
\usage{
fusion_multifused(standardFusioncalls, limitMultiFused = 3)
}
\arguments{
\item{standardFusioncalls}{A dataframe from star fusion or arriba standardized to run through the filtering steps}

\item{limitMultiFused}{Integer to identify a limit of times a gene can be fused per sample}
}
\value{
Fusions where gene partner(s) is multifused per sample
}
\description{
Function to identify fusions called by at least n callers
}
\examples{
out_annofuse <- system.file("extdata", "PutativeDriverAnnoFuse.tsv", package = "annoFuse")
sfc <- read.delim(out_annofuse)
sfc_multifused <- fusion_multifused(sfc, limitMultiFused = 2)
}
