% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusion_standardization.R
\name{fusion_standardization}
\alias{fusion_standardization}
\title{Standardizes fusion calls}
\usage{
fusion_standardization(fusion_calls, caller = c("STARFUSION", "ARRIBA"))
}
\arguments{
\item{fusion_calls}{A dataframe from star fusion or arriba (more callers to be added)}

\item{caller}{string options STARfusion/arriba}
}
\value{
Standardized fusion calls ready for filtering
}
\description{
Various fusion callers have different formats that make aggregating and filtering data difficult.
By standardizing fusion callers output we capture the required columns which we use for downstream
analysis
}
\examples{
# TODOTODO
}
