% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusionMultifused.R
\name{fusionMultifused}
\alias{fusionMultifused}
\title{Function to identify fusions called by at least n callers}
\usage{
fusionMultifused(standardFusioncalls = standardFusioncalls,
  limitMultiFused = limitMultiFused)
}
\arguments{
\item{standardFusioncalls}{A dataframe from star fusion or arriba standardized to run through the filtering steps}

\item{limitMultifused}{Integer to identify a limit of times a gene can be fused per sample}
}
\value{
Fusions where gene partner(s) is multifused per sample
}
\description{
Function to identify fusions called by at least n callers
}
