% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_exons_tracks_gtf.R
\name{get_exons_tracks_gtf}
\alias{get_exons_tracks_gtf}
\title{Function gather exon locations per gene to plot and annotate breakpoints}
\usage{
get_exons_tracks_gtf(genome = "hg38", gtf_path)
}
\arguments{
\item{genome}{human genome version (hg38 and hg19 supported)}

\item{gtf}{gencode gtf file path,  gencode.v27.primary_assembly.annotation.gtf.gz can be downloaded from "ftp://ftp.ebi.ac.uk/pub/databases/gencode/Gencode_human/release_27/gencode.v27.primary_assembly.annotation.gtf.gz"}
}
\value{
exon location dataframe
}
\description{
Function gather exon locations per gene to plot and annotate breakpoints
}
\examples{
library(EnsDb.Hsapiens.v86)
out_annofuse <- system.file("extdata", "PutativeDriverAnnoFuse.tsv", package = "annoFuse")
sfc <- read.delim(out_annofuse)
# using hg38 gencode gtf to get exon tracks
download.file("ftp://ftp.ebi.ac.uk/pub/databases/gencode/Gencode_human/release_27/gencode.v27.primary_assembly.annotation.gtf.gz","gencode.v27.primary_assembly.annotation.gtf.gz")
exons <- get_exons_tracks_gtf(gtf_path = file.path("gencode.v27.primary_assembly.annotation.gtf.gz"))
\dontrun{
ensembl <- biomaRt::useMart(biomart = "ENSEMBL_MART_ENSEMBL",dataset = "hsapiens_gene_ensembl",host="ensembl.org")
download.file("http://hgdownload.soe.ucsc.edu/goldenPath/hg38/database/pfamDesc.txt.gz","pfamDesc.txt.gz")
download.file("http://hgdownload.soe.ucsc.edu/goldenPath/hg38/database/ucscGenePfam.txt.gz","ucscGenePfam.txt.gz")
bioMartDataPfam <- get_biomart_pfam_merge(ensembl = ensembl,pfamDesc_path = "pfamDesc.txt.gz",ucscGenePfam_path = "ucscGenePfam.txt.gz")
domainDataFrame <- get_Pfam_domain(standardFusioncalls = sfc, 
                                   bioMartDataPfam = bioMartDataPfam, 
                                   keepPartialAnno = TRUE)
right <- plot_breakpoints(sampleid = "BS_044XZ8ST", 
                          domainDataFrame = domainDataFrame, 
                          exons = exons, 
                          geneposition = "Right", 
                          fusionname = "ANTXR1--BRAF",
                          rightBreakpoint = "7:140787584")
}                        
}
