% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupcount_fusion_calls.R
\name{groupcount_fusion_calls}
\alias{groupcount_fusion_calls}
\title{Function to identify fusions called by at least n callers}
\usage{
groupcount_fusion_calls(standardFusioncalls, group, numGroup)
}
\arguments{
\item{standardFusioncalls}{A dataframe from star fusion or arriba standardized to run through the filtering steps}

\item{group}{column name in input standard fusion calls that should be used to identify subset the sample belongs to in the cohort}

\item{numGroup}{Least number of groups that have the fusion to categories as false call in cohort}
}
\value{
Fusions found in more than n groups in cohort
}
\description{
Function to identify fusions called by at least n callers
}
\examples{
out_annofuse <- system.file("extdata", "PutativeDriverAnnoFuse_test_v16.tsv", package = "annoFuse")
sfc <- read.delim(out_annofuse)
sfc_groupcount <- groupcount_fusion_calls(sfc, group = "Sample", 1)
}
