% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_breakpoints.R
\name{plot_breakpoints}
\alias{plot_breakpoints}
\title{Function to plot breakpoints of a gene in a sample/cohort}
\usage{
plot_breakpoints(
  domainDataFrame = domainDataFrame,
  exons = exons,
  geneposition = geneposition
)
}
\arguments{
\item{domainDataFrame}{A dataframe from star fusion or arriba standardized to run through the filtering steps and get_Pfam_domain() function to get pfam domain annotation per gene}

\item{exons}{exon level information from gtf}

\item{geneposition}{Left/Right position of gene}
}
\value{
ggplot of gene and breakpoints
}
\description{
Function to plot breakpoints of a gene in a sample/cohort
}
\examples{
# TODOTODO

}
