% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_recurrent_fusions.R
\name{plot_recurrent_fusions}
\alias{plot_recurrent_fusions}
\title{Function to plot recurrent fused genes}
\usage{
plot_recurrent_fusions(
  standardFusioncalls,
  groupby,
  plotn = 20,
  countID,
  palette_rec = NULL,
  base_size = 20
)
}
\arguments{
\item{standardFusioncalls}{A dataframe from star fusion or arriba standardized to run through the filtering steps}

\item{groupby}{column name with grouping variables}

\item{plotn}{top n recurrent fusions to plot}

\item{countID}{column name to count recurrent fusions SampleID/ParticipantID/tumorID}

\item{palette_rec}{colors for grouping variables}

\item{base_size}{Numeric, size of font for plot}
}
\value{
A ggplot object containing an overview on the recurrent fusions
}
\description{
Function to plot recurrent fused genes
}
\examples{
out_annofuse <- system.file("extdata", "PutativeDriverAnnoFuse.tsv", package = "annoFuseData")
sfc <- read.delim(out_annofuse, stringsAsFactors = FALSE)
# keep only in-frame and fusions where both breakpoints are within genes
sfc <- as.data.frame(
  sfc[ which(sfc$Fusion_Type == "in-frame" & sfc$BreakpointLocation == "Genic"),])
plot_recurrent_fusions(sfc, 
                       groupby = "broad_histology", 
                       countID = "Kids_First_Participant_ID")
}
