% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_recurrent_genes.R
\name{plot_recurrent_genes}
\alias{plot_recurrent_genes}
\title{Function to plot recurrent fused genes}
\usage{
plot_recurrent_genes(
  standardFusioncalls = standardFusioncalls,
  groupby = groupby,
  plotn = plotn,
  countID = countID,
  palette_rec = NULL
)
}
\arguments{
\item{standardFusioncalls}{A dataframe from star fusion or arriba standardized to run through the filtering steps}

\item{groupby}{column name with grouping variables}

\item{plotn}{top n recurrent fusions to plot}

\item{countID}{column name to count recurrent fusions SampleID/ParticipantID/tumorID}

\item{palette_rec}{colors for grouping variables}
}
\value{
recurrent fused genes
}
\description{
Function to plot recurrent fused genes
}
\examples{
# TODOTODO

}
