% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_summary.R
\name{plot_summary}
\alias{plot_summary}
\title{Function to annotate fusion calls}
\usage{
plot_summary(standardFusioncalls, outputpdffile, groupby)
}
\arguments{
\item{standardFusioncalls}{A dataframe from star fusion or arriba standardized to run through the filtering steps}

\item{outputpdffile}{Filename to plot image - optional, if not specified, the 
plot is simply generated}

\item{groupby}{column name with grouping variables}
}
\value{
A ggplot object containing a summary of the standardized fusion calls
}
\description{
Function to annotate fusion calls
}
\examples{
out_annofuse <- system.file("extdata", "PutativeDriverAnnoFuse.tsv", package = "annoFuse")
sfc <- read.delim(out_annofuse,stringsAsFactors = FALSE)
plot_summary(sfc, groupby = "Fusion_Type",outputpdffile = "out.pdf")
}
