% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_breakpoints..R
\name{plotBreakpoints}
\alias{plotBreakpoints}
\title{Function to plot breakpoints of a gene in a sample/cohort}
\usage{
plotBreakpoints(
  domainDataFrame = domainDataFrame,
  exons = exons,
  geneposition = geneposition
)
}
\arguments{
\item{domainDataFrame}{A dataframe from star fusion or arriba standardized to run through the filtering steps and getPfamDomain() function to get pfam domain annotation per gene}

\item{exons}{exon level information from gtf}

\item{geneposition}{Left/Right position of gene}
}
\value{
ggplot of gene and breakpoints
}
\description{
Function to plot breakpoints of a gene in a sample/cohort
}
