% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPercentilePerGene.R
\name{plotPercentilePerGene}
\alias{plotPercentilePerGene}
\title{Function to plot expression as percentile}
\usage{
plotPercentilePerGene(expressiongene = expressiongene,
  geneSymbol = geneSymbol, standardFusionCalls = standardFusionCalls)
}
\arguments{
\item{expressiongene}{matrix with 1 row with expression for gene and N columns for samples to be used for analysis}

\item{geneSymbol}{geneSymbol}

\item{standardFusionCalls}{standardized fusion calls}
}
\value{
ggplot grob for plotting median and each value per fusion as x axis
}
\description{
Function to plot expression as percentile
}
