% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_recurrent_genes.R
\name{plotRecurrentGenes}
\alias{plotRecurrentGenes}
\title{Function to plot recurrent fused genes}
\usage{
plotRecurrentGenes(standardFusioncalls = standardFusioncalls,
  groupby = groupby, plotn = plotn, countID = countID)
}
\arguments{
\item{standardFusioncalls}{A dataframe from star fusion or arriba standardized to run through the filtering steps}

\item{groupby}{column name with grouping variables}

\item{plotn}{top n recurrent fusions to plot}

\item{countID}{column name to count recurrent fusions SampleID/ParticipantID/tumorID}
}
\value{
recurrent fused genes
}
\description{
Function to plot recurrent fused genes
}
