% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annoFuse-utils.R
\name{read_arriba_calls}
\alias{read_arriba_calls}
\title{Read in fusion calls from Arriba}
\usage{
read_arriba_calls(arriba_calls)
}
\arguments{
\item{arriba_calls}{TODO}
}
\value{
A data.frame object with correct column specifications
}
\description{
Read in fusion calls from Arriba
}
\examples{
# TODO
}
