% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annoFuse-utils.R
\name{read_starfusion_calls}
\alias{read_starfusion_calls}
\title{Read in fusion calls from STAR Fusion}
\usage{
read_starfusion_calls(starfusion_calls)
}
\arguments{
\item{starfusion_calls}{TODO}
}
\value{
A data.frame object with correct column specifications
}
\description{
Read in fusion calls from STAR Fusion
}
\examples{
# TODO
}
