% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_fuse.R
\name{report_fuse}
\alias{report_fuse}
\title{report_fuse}
\usage{
report_fuse(
  out_annofuse,
  project_id = "projectID",
  input_rmd = NULL,
  output_file = "my_first_annoFuse_report.html",
  output_dir = tempdir(),
  output_format = NULL,
  force_overwrite = FALSE,
  knitr_show_progress = FALSE,
  open_after_creating = TRUE,
  ...
)
}
\arguments{
\item{out_annofuse}{The character string specifying the location of the file
output by the annoFuse pipeline.}

\item{project_id}{A character string, which can be considered as an identifier
for the set/session, and will be e.g. used in the title of the report created
via [report_fuse()]}

\item{input_rmd}{Character string with the path to the RMarkdown (.Rmd) file
that will be used as the template for generating the report. Defaults to NULL,
which will then use the one provided with the `annoFuse` package.}

\item{output_file}{Character string, specifying the file name of the output
report. The file name extension must be either `.html` or `.pdf`, and consistent
with the value of `output_format`.}

\item{output_dir}{Character, defining the path to the output directory where
the report will be generated. Defaults to the temp directory (`tempdir()`).}

\item{output_format}{The format of the output report. Either `html_document`
or `pdf_document`. The file name extension of `output_file` must be consistent
with this choice. Can also be left empty and determined accordingly.}

\item{force_overwrite}{Logical, whether to force overwrite an existing report
with the same name in the output directory. Defaults to FALSE.}

\item{knitr_show_progress}{Logical, whether to display the progress of `knitr`
while generating the report. Defaults to FALSE.}

\item{open_after_creating}{Logical, whether to open the report in the default
browser after being generated. Defaults to TRUE.}

\item{...}{Other arguments that will be passed to [rmarkdown::render()].}
}
\value{
Generates a fully fledged report in the `output_dir` directory, called
`output_file` and returns (invisibly) the name of the generated report.
}
\description{
report_fuse
}
\details{
When `report_fuse` is called, a RMarkdown template file could be copied
into the output directory, and [rmarkdown::render()] will be called to
generate the final report.

As a default template, `report_fuse` uses the one delivered together with the
`annoFuse` package. Experienced users can take that as a starting point to further
edit and customize.
}
\examples{

out_annofuse <- system.file("extdata", "PutativeDriverAnnoFuse.tsv", package = "annoFuse")
\dontrun{
report_fuse(out_annofuse = out_annofuse)
}
}
