% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplecount_fusion_calls.R
\name{samplecount_fusion_calls}
\alias{samplecount_fusion_calls}
\title{Function to identify fusions called in n samples}
\usage{
samplecount_fusion_calls(standardFusioncalls, numSample = 1, group)
}
\arguments{
\item{standardFusioncalls}{A dataframe from star fusion or arriba standardized to run through the filtering steps}

\item{numSample}{Least number of samples per group that have the fusion. Defaults to 1.}

\item{group}{column name for grouping variable}
}
\value{
Fusions found in atleast n samples
}
\description{
Function to identify fusions called in n samples
}
\examples{
out_annofuse <- system.file("extdata", "PutativeDriverAnnoFuse.tsv", package = "annoFuseData")
sfc <- read.delim(out_annofuse)
samplecount_fusion_calls(sfc, group = "Kids_First_Participant_ID")
}
