% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusion_standardization.R
\name{shape_output}
\alias{shape_output}
\title{function used by STARFUSION and ARRIBA to shape the final output columns as required}
\usage{
shape_output(fusion_calls)
}
\arguments{
\item{fusion_calls}{A dataframe used for star fusion or arriba method  (more callers to be added)}
}
\value{
processed dataframe in the final desired format
}
\description{
function used by STARFUSION and ARRIBA to shape the final output columns as required
}
\author{
Saksham Phul (phuls@chop.edu)
}
