% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_publication.R
\name{theme_publication}
\alias{theme_publication}
\title{theme_publication for plots
Publication quality images for summary plots}
\usage{
theme_publication(base_size = 12, base_family = "Helvetica")
}
\arguments{
\item{base_size}{size of font for plot default (defaults to 12)}

\item{base_family}{font to be used default (defaults to "Helvetica")}
}
\value{
An object of class \code{\link{theme}} for styling the plots.
}
\description{
theme_publication for plots
Publication quality images for summary plots
}
\examples{
library("ggplot2")
p <- ggplot(cars, aes(speed, dist)) + geom_point()
p
p + theme_publication(base_size = 24)
}
