% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annoFuse_single_sample.R
\name{annoFuse_single_sample}
\alias{annoFuse_single_sample}
\title{Single Sample use for annoFuse}
\usage{
annoFuse_single_sample(
  fusionfileArriba = fusionfileArriba,
  fusionfileStarFusion = fusionfileStarFusion,
  expressionFile = NULL,
  expressionFilter = 1,
  tumorID = "tumorID",
  readingFrameFilter = "in-frame|frameshift|other",
  readthroughFilter = FALSE,
  artifactFilter = "GTEx_Recurrent|DGD_PARALOGS|Normal|BodyMap|ConjoinG",
  junctionReadCountFilter = 1,
  spanningFragCountFilter = 10
)
}
\arguments{
\item{fusionfileArriba}{A dataframe from arriba fusion caller}

\item{fusionfileStarFusion}{A dataframe from starfusion caller}

\item{expressionFile}{Expression matrix for samples used in cohort for fusion calls}

\item{expressionFilter}{FPKM/TPM threshold for not expressed}

\item{tumorID}{Sample name to be used}

\item{readingFrameFilter}{A regex to capture readingframe (eg. in-frame|frameshift|other)}

\item{readthroughFilter}{Boolean for filtering readthroughs}

\item{artifactFilter}{A red flag filter from Annotation ; in OpenPBTA annotation is from FusionAnnotator column "annots"}

\item{junctionReadCountFilter}{An integer threshold for JunctionReadCount}

\item{spanningFragCountFilter}{An integer threshold for (SpanningFragCount - JunctionReadCount)}
}
\value{
Standardized fusion calls annotated with gene list and fusion list provided in reference folder
}
\description{
Performs artifact filter to remove readthroughs,red flags and performs expression filtering with user provided expression Matrix and expression threshold
}
\examples{
# TODOTODO
}
