% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_fusion_calls.R
\name{annotate_fusion_calls}
\alias{annotate_fusion_calls}
\title{Function to annotate fusion calls}
\usage{
annotate_fusion_calls(
  standardFusioncalls,
  geneListReferenceDataTab,
  fusionReferenceDataTab
)
}
\arguments{
\item{standardFusioncalls}{A dataframe from star fusion or arriba standardized to run through the filtering steps}

\item{geneListReferenceDataTab}{A dataframe with column 1 as GeneName 2 source file 3 type; collapse to summarize type}

\item{fusionReferenceDataTab}{A dataframe with column 1 as FusionName 2 source file 3 type; collapse to summarize type}
}
\value{
Standardized fusion calls annotated with gene list and fusion list provided in reference folder
}
\description{
Function to annotate fusion calls
}
\examples{
# standardize
fusionfileArriba = read.delim(system.file("extdata", "arriba_example.tsv", package = "annoFuse"),stringsAsFactors = FALSE)
fusionfileStarFusion = read.delim(system.file("extdata", "starfusion_example.tsv", package = "annoFuse"),stringsAsFactors = FALSE)
formattedArriba = fusion_standardization(fusionfileArriba,caller="ARRIBA",tumorID = "tumorID")
formattedStarFusion = fusion_standardization(fusionfileStarFusion,caller="STARFUSION",tumorID = "tumorID")
# merge standardized fusion calls
standardFusioncalls <- rbind(formattedStarFusion, formattedArriba) \%>\% as.data.frame()
fusionQCFiltered <- fusion_filtering_QC(standardFusioncalls = standardFusioncalls, 
                    readingFrameFilter = "in-frame|frameshift|other",
                    artifactFilter = "GTEx_Recurrent|DGD_PARALOGS|Normal|BodyMap|ConjoinG",
                    junctionReadCountFilter = 1,
                    spanningFragCountFilter = 10,
                    readthroughFilter = TRUE)
# annotated from gene and fusion refrence list
# read in gene and fusion reference tab
geneListReferenceDataTab <- read.delim(system.file("extdata", "genelistreference.txt", package = "annoFuse"), stringsAsFactors = FALSE)
# column 1 as FusionName 2 source file 3 type; collapse to summarize type
fusionReferenceDataTab <- read.delim(system.file("extdata", "fusionreference.txt", package = "annoFuse"), stringsAsFactors = FALSE)
filteredFusionAnnotated <- annotate_fusion_calls(standardFusioncalls = fusionQCFiltered, geneListReferenceDataTab = geneListReferenceDataTab, fusionReferenceDataTab = fusionReferenceDataTab)                                   
}
