% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusion_driver.R
\name{fusion_driver}
\alias{fusion_driver}
\title{Filter standardized fusion calls for driver fusions}
\usage{
fusion_driver(
  standardFusioncalls,
  annotated = TRUE,
  geneListReferenceDataTab,
  fusionReferenceDataTab,
  checkDomainStatus = FALSE,
  domainsToCheck
)
}
\arguments{
\item{standardFusioncalls}{A dataframe from star fusion or arriba (more callers to be added)}

\item{annotated}{Boolean if annotated}

\item{geneListReferenceDataTab}{A dataframe with column 1 as GeneName 2 source file 3 type; collapse to summarize type}

\item{fusionReferenceDataTab}{A dataframe with column 1 as FusionName 2 source file 3 type; collapse to summarize type}

\item{checkDomainStatus}{Logical value to check if domain status in fused gene for given domansToCheck, default to FALSE}

\item{domainsToCheck}{pfamID to check for retention status, the IDs can be found here http://hgdownload.soe.ucsc.edu/goldenPath/hg38/database/pfamDesc.txt.gz}
}
\value{
Putative Driver standardized fusion calls annotated with gene list and fusion list provided in reference folder.If checkDomainStatus =TRUE and domain retention status for given pfamID is also provided along with the gene location corresponding to the domain retention status
}
\description{
If standardized fusion calls are annotated using the geneListReferenceDataTab and fusionReferenceDataTab filters out fusion calls where partner genes are not annotated.
If standardized fusion is not annotated it will be annotated with geneListReferenceDataTab and fusionReferenceDataTab provided
}
\examples{
out_annofuse <- system.file("extdata", "PutativeDriverAnnoFuse.tsv", package = "annoFuse")
sfc <- read.delim(out_annofuse)
geneListReferenceDataTab <- read.delim(
  system.file("extdata", "genelistreference.txt", package="annoFuse"), stringsAsFactors = FALSE)
fusionReferenceDataTab <- read.delim(
  system.file("extdata", "fusionreference.txt", package="annoFuse"), stringsAsFactors = FALSE)
  
bioMartDataPfam <- readRDS(system.file("extdata", "pfamDataBioMart.RDS", package = "annoFuse"))
kinaseid<-unique(bioMartDataPfam$pfam_id[grep("kinase",bioMartDataPfam$NAME)] )
fusion_driver_df <- fusion_driver(sfc,
                                  annotated = TRUE,
                                  geneListReferenceDataTab = geneListReferenceDataTab,
                                  fusionReferenceDataTab = fusionReferenceDataTab,
                                  checkDomainStatus=FALSE,
                                  domainsToCheck=kinaseid)

}
