% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusion_standardization.R
\name{fusion_standardization}
\alias{fusion_standardization}
\title{K. S. Gaonkar 2019}
\usage{
fusion_standardization(fusion_calls = fusion_calls, caller = caller)
}
\arguments{
\item{fusion_calls}{A dataframe from star fusion or arriba (more callers to be added)}

\item{caller}{string options STARfusion/arriba}
}
\value{
Standardized fusion calls ready for filtering
}
\description{
K. S. Gaonkar 2019
}
