% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_exons_tracks_gtf.R
\name{get_exons_tracks_gtf}
\alias{get_exons_tracks_gtf}
\title{Function gather exon locations per gene to plot and annotate breakpoints}
\usage{
get_exons_tracks_gtf(genome = "hg38", gtf_path)
}
\arguments{
\item{genome}{human genome version (hg38 and hg19 supported)}

\item{gtf_path}{gencode gtf file path,  gencode.v27.primary_assembly.annotation.gtf.gz can be downloaded from "ftp://ftp.ebi.ac.uk/pub/databases/gencode/Gencode_human/release_27/gencode.v27.primary_assembly.annotation.gtf.gz"}
}
\value{
exon location dataframe
}
\description{
Function gather exon locations per gene to plot and annotate breakpoints
}
\examples{
library(EnsDb.Hsapiens.v86)

\dontrun{
# using hg38 gencode gtf to get exon tracks
download.file("ftp://ftp.ebi.ac.uk/pub/databases/gencode/Gencode_human/release_27/gencode.v27.primary_assembly.annotation.gtf.gz", "gencode.v27.primary_assembly.annotation.gtf.gz")
exons <- get_exons_tracks_gtf(gtf_path = file.path("gencode.v27.primary_assembly.annotation.gtf.gz"))
}

}
