% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_Pfam_domain.R
\name{get_Pfam_domain}
\alias{get_Pfam_domain}
\title{Function to annotate fusion calls with domain terms}
\usage{
get_Pfam_domain(standardFusioncalls, bioMartDataPfam, keepPartialAnno = FALSE)
}
\arguments{
\item{standardFusioncalls}{A dataframe from star fusion or arriba standardized to run through the filtering steps}

\item{bioMartDataPfam}{A dataframe with gene and domain coordinate information with chromosome_name,gene_start,gene_end,domain_chr,domain_start,domain_end,hgnc_symbol}

\item{keepPartialAnno}{TRUE or FALSE to keep partial status; defaults to FALSE}
}
\value{
Standardized fusion calls as list Gene1A and Gene1B annotated with domain terms and chromosome location; retained and not retained,optionally partially retained
}
\description{
Function to annotate fusion calls with domain terms
}
\examples{
out_annofuse <- system.file("extdata", "PutativeDriverAnnoFuse.tsv", package = "annoFuse")
sfc <- read.delim(out_annofuse)
bioMartDataPfam <- readRDS(system.file("extdata", "pfamDataBioMart.RDS", package = "annoFuse"))
domain_list_df <- get_Pfam_domain(standardFusioncalls = sfc,bioMartDataPfam = bioMartDataPfam)
}
