% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_recurrent_genes.R
\name{plot_recurrent_genes}
\alias{plot_recurrent_genes}
\title{Function to plot recurrent fused genes}
\usage{
plot_recurrent_genes(
  standardFusioncalls,
  groupby,
  plotn = 20,
  countID,
  palette_rec = NULL,
  base_size = 12
)
}
\arguments{
\item{standardFusioncalls}{A dataframe from star fusion or arriba standardized to run through the filtering steps}

\item{groupby}{column name with grouping variables}

\item{plotn}{top n recurrent fusions to plot}

\item{countID}{column name to count recurrent fusions SampleID/ParticipantID/tumorID}

\item{palette_rec}{colors for grouping variables}

\item{base_size}{Numeric, size of font for plot}
}
\value{
A ggplot object containing an overview on the recurrent fused genes
}
\description{
Function to plot recurrent fused genes
}
\examples{
out_annofuse <- system.file("extdata", "PutativeDriverAnnoFuse_test_v14.tsv",
  package = "annoFuse"
)
sfc <- read.delim(out_annofuse)
clinical <- read.delim(
  system.file("extdata", "pbta-histologies.tsv", package = "annoFuse")
)
# Select only in-frame and frameshift
library("dplyr")
sfc <- sfc \%>\%
  dplyr::filter(Fusion_Type \%in\% c("in-frame", "frameshift"))
sfc <- merge(sfc, clinical[, c("Kids_First_Biospecimen_ID", "broad_histology")],
  by.x = "Sample", by.y = "Kids_First_Biospecimen_ID"
)
# Remove Benign tumor fusions
sfc <- sfc[-which(sfc$FusionName \%in\%
  unique(sfc[which(sfc$broad_histology == "Benign tumor"), "FusionName"])), ]
# Remove GeneA == GeneB
sfc <- sfc[-which(sfc$Gene1A == sfc$Gene1B |
  sfc$Gene1A == sfc$Gene2B |
  sfc$Gene1B == sfc$Gene2A), ]
# Remove intergenic fusions
sfc <- sfc[-grep("/", sfc$FusionName), ]
plot_recurrent_genes(sfc, groupby = "Fusion_Type", countID = "Sample")
}
