% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_recurrent_genes.R
\name{plot_recurrent_genes}
\alias{plot_recurrent_genes}
\title{Function to plot recurrent fused genes}
\usage{
plot_recurrent_genes(
  standardFusioncalls,
  groupby,
  plotn = 20,
  countID,
  palette_rec = NULL
)
}
\arguments{
\item{standardFusioncalls}{A dataframe from star fusion or arriba standardized to run through the filtering steps}

\item{groupby}{column name with grouping variables}

\item{plotn}{top n recurrent fusions to plot}

\item{countID}{column name to count recurrent fusions SampleID/ParticipantID/tumorID}

\item{palette_rec}{colors for grouping variables}
}
\value{
A ggplot object containing an overview on the recurrent fused genes
}
\description{
Function to plot recurrent fused genes
}
\examples{
out_annofuse <- system.file("extdata", "PutativeDriverAnnoFuse.tsv", package = "annoFuse")
sfc <- read.delim(out_annofuse)
library(tidyverse)
sfc <- sfc \%>\%
   dplyr::filter(Fusion_Type=="in-frame",BreakpointLocation=="Genic") \%>\% 
   as.data.frame()    
plot_recurrent_genes(sfc, groupby = "broad_histology", countID = "Kids_First_Participant_ID")
}
