% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_summary.R
\name{plot_summary}
\alias{plot_summary}
\title{Function to annotate fusion calls}
\usage{
plot_summary(
  standardFusionCalls = standardFusionCalls,
  outputpdffile = outputpdffile,
  groupby = groupby
)
}
\arguments{
\item{standardFusionCalls}{A dataframe from star fusion or arriba standardized to run through the filtering steps}

\item{outputpdffile}{Filename to plot image}

\item{groupby}{column name with grouping variables}
}
\value{
summary pdf
}
\description{
Function to annotate fusion calls
}
\examples{
# TODOTODO

}
