% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_summary.R
\name{plot_summary}
\alias{plot_summary}
\title{Function to annotate fusion calls}
\usage{
plot_summary(standardFusioncalls, groupby, outputpdffile, base_size = 20)
}
\arguments{
\item{standardFusioncalls}{A dataframe from star fusion or arriba standardized to run through the filtering steps}

\item{groupby}{column name with grouping variables}

\item{outputpdffile}{Filename to plot image - optional, if not specified, the 
plot is simply generated}

\item{base_size}{size for text in plots, defalt 20 for pdf outputs}
}
\value{
A ggplot object containing a summary of the standardized fusion calls
}
\description{
Function to annotate fusion calls
}
\examples{
out_annofuse <- system.file("extdata", "PutativeDriverAnnoFuse.tsv", package = "annoFuse")
sfc <- read.delim(out_annofuse, stringsAsFactors = FALSE)
plot_summary(sfc, groupby = "Fusion_Type")
}
