% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSummary.R
\name{plotSummary}
\alias{plotSummary}
\title{Function to annotate fusion calls}
\usage{
plotSummary(
  standardFusionCalls = standardFusionCalls,
  outputpdffile = outputpdffile,
  groupby = groupby
)
}
\arguments{
\item{standardFusionCalls}{A dataframe from star fusion or arriba standardized to run through the filtering steps}

\item{outputpdffile}{Filename to plot image}

\item{groupby}{column name with grouping variables}
}
\value{
summary pdf
}
\description{
Function to annotate fusion calls
}
