% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annoFuse-utils.R
\name{read_starfusion_calls}
\alias{read_starfusion_calls}
\title{Read in fusion calls from STAR-Fusion 1.5.0}
\usage{
read_starfusion_calls(starfusion_calls)
}
\arguments{
\item{starfusion_calls}{Please refer to software documenation [star-fusion.fusion_predictions.tsv] https://github.com/STAR-Fusion/STAR-Fusion/wiki#output-from-star-fusion}
}
\value{
A data.frame object with correct column specifications
}
\description{
Read in fusion calls from STAR-Fusion 1.5.0
}
\examples{
fusionfileStarFusion <- read_starfusion_calls(
  system.file("extdata", "starfusion_example.tsv", package = "annoFuseData"))
                           
}
