% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplecount_fusion_calls.R
\name{samplecount_fusion_calls}
\alias{samplecount_fusion_calls}
\title{Function to identify fusions called in n samples}
\usage{
samplecount_fusion_calls(
  standardFusioncalls = standardFusioncalls,
  numSample = numSample,
  group = group
)
}
\arguments{
\item{standardFusioncalls}{A dataframe from star fusion or arriba standardized to run through the filtering steps}

\item{numSample}{Least number of samples per group that have the fusion}

\item{group}{column name for grouping variable}
}
\value{
Fusions found in atleast n samples
}
\description{
Function to identify fusions called in n samples
}
\examples{
# TODOTODO

}
