% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_fuse.R
\name{shiny_fuse}
\alias{shiny_fuse}
\title{shiny_fuse}
\usage{
shiny_fuse(out_annofuse)
}
\arguments{
\item{out_annofuse}{TODO}
}
\value{
TODO
}
\description{
TODO
}
\examples{
out_annofuse <- "/Users/fede/Development/annoFuse/PutativeDriverAnnoFuse_test_v14.tsv"
if (interactive()) {
  shiny_fuse(out_annofuse)
}
}
