% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_publication.R
\name{theme_Publication}
\alias{theme_Publication}
\title{theme_publication for plots}
\usage{
theme_Publication(base_size = 15, base_family = "Helvetica")
}
\arguments{
\item{base_size}{size of font for plot default "15"}

\item{base_family}{font to be used default "Helvetica"
Publication quality images for summary plots}
}
\description{
theme_publication for plots
}
