% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_publication.R
\name{theme_Publication}
\alias{theme_Publication}
\title{theme_publication for plots
Publication quality images for summary plots}
\usage{
theme_Publication(base_size = 12, base_family = "Helvetica")
}
\arguments{
\item{base_size}{size of font for plot default "15"}

\item{base_family}{font to be used default "Helvetica"}
}
\value{
An object of class \code{\link{theme}} for styling the plots.
}
\description{
theme_publication for plots
Publication quality images for summary plots
}
