% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zscored_annotation.R
\name{zscored_annotation}
\alias{zscored_annotation}
\title{Annotates standardizes fusion calls calls with zscored expression value from either GTEx/cohort}
\usage{
zscored_annotation(
  standardFusionCalls = standardFusionCalls,
  zscoreFilter = zscoreFilter,
  saveZscoredMatrix = saveZscoredMatrix,
  normData = normData,
  expressionMatrix = expressionMatrix
)
}
\arguments{
\item{standardFusionCalls}{Annotates standardizes fusion calls from callers STARfusion| Arriba or QC filtered fusion}

\item{zscoreFilter}{Zscore value to use as threshold for annotation of differential expression}

\item{saveZscoredMatrix}{File to save zscored matrix calculated for the normalized data and expression matrix}

\item{normData}{normalizing expression dataset to calculate zscore}

\item{expressionMatrix}{Expression matrix associated with the fusion calls}
}
\value{
expression_annotated_fusions is a standardized fusion call set with standard
}
\description{
The input should have the following standardized
columns to run through this GTEx/cohort normalization function
"Sample"  Unique SampleIDs used in your RNAseq dataset
"FusionName" GeneA--GeneB ,or if fusion is intergenic then Gene1A/Gene2A--GeneB
}
\examples{
# TODOTODO
}
